requirements.txt
aiogram==2.25.1
SQLAlchemy==1.4.48
aiosqlite==0.19.0
python-dotenv
requests
meilisearch